feat: Update chatbot configuration with new chatflowid and apiHost
This commit updates the chatbot configuration in index.html to use a new chatflowid and apiHost. The chatflowid is changed to '46b63d40-cc1e-47c1-84f0-8573a809fb6f' and the apiHost is changed to 'https://reactjsopenai.azurewebsites.net'. These changes are made to ensure the chatbot is connected to the correct chatflow and API endpoint.